+++ test-tools/issuer-front-end/CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Unreleased changes
 
+## 1.0.5
+
+- Add support for optional attributes.
+- Credential schema is only loaded from the contract instance.
+
 ## 1.0.4
 
 - Add test case to restore credential.